Move tests.utils module to tests.helpers.io per Alex's suggestion
